Moved testBook.xlsx from the project root to Raw\Excel Files\test
Fixed an issue which caused the program to crash if "prefer" was used instead of "pref" as a preferred date constraint prefix.

Fixed the immidiate problem with the new Parser.java test class which was a mix between not having the configuration file loaded and not having a valid conference set in testBook.xlsx. Finding and fixing the issue with the test file led me to find Issue #20, and I subsequently marked those locations with the FIXME tag for whoever chooses to take on that bug.