Different numbers of timeslots on different days is looking painful to implement.
My basic plan was to have a "standardDates" array which would be initialized first, and then
serve as the blueprint for every DateConstraint's dates array. Mostly, it's looking like before/after
constraints are going to be hard to implement if there can be a different number of timeslots on
every day.

#1176539