Merge of commit Matthew made.
Refactored the CourtScheduleInfo code so that the variable names are more descriptive.

Made the output path an optional field in the config.ini -- if it's not set, it uses the parent directory of the input file as the output folder.

Modified the README to include a description on what the expected input filename is.